Create kmeans.m
Modification of the vanilla 'kmeans' function that comes with GNU Octave 'statistics' package, to support initialization of centroids as specific indices of the data to be used. The array of indices must be of same length as the number of centroids. A demo is provided at the end for usage.

This also includes a temporary fix to errors when the function is run with extended params but without the "emptyaction" property, as described here:
https://lists.gnu.org/archive/html/octave-bug-tracker/2014-04/msg00293.html